Begin separating the Xlib API from the perl OO layer
I thought it would be good to allow developers direct
access to the Xlib API calls, without forcing them to
use the object layer.  On the same note, I want to make
the object layer smarter, which could get in the way
of someone who wants the lowest level access.   I also
want to make it extensible, by blessing hashes instead
of blessing C pointers, and people might want to avoid
that overhead.

I started by creating a X11::Xlib::Display object that
wil take over the role of doing smart things with the
connection, and letting X11::Xlib just be the namespace
for the Xlib API.  For backward compatibility, the new
Display object is a subclass of X11::Xlib, and the
'new' method of X11::Xlib redirects to the Display
object.  The original test cases still pass.

I enhanced the typemap to accept both Xlib and Display
objects as input for Display*, and allow both integers
and objects as input for XIDs.  It now always returns
the simplest form of as Output with the idea that the
OO layer can wrap that in an appropriate object.

This also gives developers the ability to subclass the
behavior when extending the Window or Display objects.