New methods for error handling
I plan to add a feature where errors can be trapped and
directed to callbacks on the Window or Display object,
and which allow the user to do a graceful recovery after
a fatal X error, but keep the user safe by invalidating
all the connections.  Feature is almost done.