Cleaned up internals, and fixed error trapping
I gave up on XS tricks to mutate the object, and insead just
allocate a small struct to hold my additional status bits.
Made the code a lot cleaner.

Then, I fixed the error handling so that it behaves like I
planned, with all handles becoming invalid if Xlib hits a fatal
error, and removing themselves from the connection cache when
closed, and giving helpful error messages when padding a dead
or closed handle to an Xlib API function.