Close conn on X11::Xlib insead of Display obj
I had already updated the documentation to say that X11::Xlib
instances automatically close the connection, but had not
updated the code to match.  Now the destructor of X11::Xlib
checks whether the connection is open and is not foreign, and
calls XCloseDisplay if needed.   The X11::Xlib::Display object
no longer needs a destructor.