Handle coercion to struct different for lvalue than rvalue
This was necesary to coerce lvalue XEvent objects to the correct
type, but not alter the objects being passed as rvalues to other
functions.

Overall, the behavior is much better now, with appropriate
exceptions for operations that don't make sense, like passing
undef for an rvalue, passing rvalue objects of the wrong type,
or using hashrefs for an lvalue.