More rigorous coercion for structs and Display wrapping
I finally enumerated the exact things that are
allowed to be coerced to Struct objects, differing
for lvalues and rvalues.  Then added test cases.

Also, found some bugs in the creation of Display
object wrappers, and added a new 'create' parameter
to avoid creating new instances in cases where it
is almost certainly an error.