Testcase for XComposite using Xephyr
New plan is to run the test cases in a sandbox when
possible.  I added a class that (will eventually)
tests for different known X11 servers, and then picks
one and runs it for the duration of the test.

Then, wrote test case for some Xcomposite calls.

Then fixed a bug in XFixesCreateRegion.