Fix XErrorEvent member of XEvent
XErrorEvent is part of the XEvent union, but breaks field
order with XAnyEvent, throwing a wrench into my optimizations.
Also, I forgot that I omitted it from the generated XS, which
resulted in a completely broken error handler.

Also, it looks like type 0 is XErrorEvent, although it
doesn't have a constant like the rest.