Pass object reference to _unpack C functions
In some cases, the _unpack function needs to know what the
current Display* is, and the struct does not always include
this information.  By passing the object ref as a third
argument, the ->display attribute will always be available
even for the current case where it is implemented as an
inside-out attribute that comes from a package global.

This would change the public API though, so I created a new
name for the functions, and added wrappers to continue to
expose the old ones.