Use correct length for the XGetWindowProperty/XChangeProperty data.
From the manual for XGetWindowProperty:

"If the returned format is 8, the returned data is represented as a
char array. If the returned format is 16, the returned data is
represented as a short array and should be cast to that type to obtain
the elements. If the returned format is 32, the returned data is
represented as a long array and should be cast to that type to obtain
the elements."

From the manual for XChangeProperty:

"If the specified format is 8, the property data must be a char array.
If the specified format is 16, the property data must be a short
array. If the specified format is 32, the property data must be a long
array."

Use the size of the corresponding data type (char/short/long) when
calculating the length of the data.