Stringify ATOMs to just the name
I was using "123=NAME" format as a way of showing that the variable
could be used in numeric context, however the ATOMs don't get consistent
values so this made it un-usable in hash keys by known names.  I decided
to just let the numeric aspect be hidden.

Also, fixed a bug where is_an_integer(undef) caused a warning.