Fix the function signature for PerlXlib_magic_vt methods
Threaded perl passes an additional argument to the magic-vtable
methods of the thread context.  Failing to declare the macro for
this parameter causes Perl to crash any time X11 objects are
freed.

Also, fixed a large number of printf '%' warnings that would have
been a problem on some 32-bit configurations.